AssemblyLoader clean up
