+++ Metalama.Framework/src/tests/.editorconfig
@@ -1,0 +1,6 @@
+# If you need any customization of .editorconfig for tests only, this is the right place to add them.
+# The .editorconfig in the repo root is a symlink to a file that is common to all repos.
+
+[*.cs]
+dotnet_diagnostic.IDE0009.severity = none
+
